Revert "fix(designer): Canvas properly centered on load (#5226)"
This reverts commit 13b018a4d7123396cce9037a950422bdf45a2941.
